move validation into common file for client use
